refactor: create separate NotificationConfig entity
