+++ src/domain/entities/NotificationConfig.ts
@@ -1,0 +1,8 @@
+import { NamedRef } from "./Ref";
+
+export type NotificationConfig = {
+    permissions: {
+        users: NamedRef[];
+        userGroups: NamedRef[];
+    };
+};
