+++ src/domain/repositories/NotificationConfigRepository.ts
@@ -1,5 +1,5 @@
 import { FutureData } from "../types/Future";
-import { NotificationConfig } from "../entities/Notification";
+import { NotificationConfig } from "../entities/NotificationConfig";
 
 export interface NotificationConfigRepository {
     get(): FutureData<NotificationConfig>;
